source/requirements-linux.txt
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.17.4
pkg-resources==0.0.0
PySide2==5.13.2
python-dateutil==2.8.1
scipy==1.3.3
shiboken2==5.13.2
six==1.13.0
